Append cluster type suffix to kube-state-metrics (#10062)
* Prepare KSM component for cluster type specific name suffix

In the local setup, seed cluster and garden-runtime cluster Pods are
running in a shared cluster. Both components share a kube-state-metrics
Deployment. This commit prepares the shared kube-state-metrics component
to be deployed with a cluster type specific suffixe (-seed and -runtime).
This will allow have cluster type specific configurations of CRD monitoring
in future.

The already existing Prometheus scrape configs here: ...

				Action:      "replace",
				Replacement: ptr.To("kube-state-metrics"),
				TargetLabel: "job",

override the job name to kube-state-metrics. This avoids breaking any
existing dashboards, federations, or rules. The individual
kube-state-metrics instances are distinguished by the component label
during scraping.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Deploy KSM with cluster type specific suffix

After preparing the share kube-state-metrics component for a name
suffix, this commit adds a `-seed` and `-runtime` suffix to the
deployments that are created by the gardenlet and garden operator.

* Adapt KSM component unit tests for name suffix

The shared KSM deployment got support for a cluster type specific suffix
like `-seed` or `-runtime`. This commit adapts the existing seed cluster
specific unit tests to add the `-seed` suffix to the component and
validates that the deployment and all related resources are created with
the correct suffix.

* Add test for garden-runtime cluster KSM

Existing unit tests were covering the deployment of kube-state-metrics for
seed and shoot clusters. This commit adds an equivalent unit test for the
garden-runtime cluster.

* Use separate ClusterRole and ClusterRoleBinding

The clusterType is ClusterTypeSeed both for the seed and the runtime
kube-state-metrics instances so the suffix "-seed" is not sufficient to
distinguish between them.

For the runtime-garden cluster, the kube-state-metrics instance is going to get
an additional role to be able to monitor the Garden resource in a future PR.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Deploy KSM to seed if it is a garden cluster

Previously, the gardenlet did not deploy the kube-state-metrics instance to the
garden namespace of the seed, if the seed cluster was a garden cluster. This was
done to avoid having two kube-state-metrics instances in the garden namespace:
one deployed by the gardener operator and the other deployed by the gardenlet,
because they used to be equivalent.

The two kube-state-metrics instances are going to be different in the near
future. Hence, both should be deployed to the garden namespace if the seed
cluster is the Garden cluster.

- The kube-state-metrics instance deployed by the gardener operator is going to
  be used to monitor the Garden resource in the runtime-garden cluster in a
  future PR, in addition to the pods in the runtime-garden cluster.

- The kube-state-metrics instance deployed by the gardenlet is going to be used
  to monitor the pods in the seed cluster, as before.

So although there is an overlap in the 2 kube-state-metrics instances (both
monitor the pods), only the runtime instance is going to monitor the Garden
resource.

In the local setup, the garden cluster is a seed cluster, so by deploying both
instances we can try and test both configurations.

Note that in real landscapes, the Garden cluster is typically not used as a seed
cluster.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Cleanup KSM without suffix

After attaching a cluster type specific suffix to the seed and garden
kube-state-metrics deployment, this change cleans up the ManagedResource
object for any un-suffixed resource.

* Fix linting errors

---------

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>