Add `HasClass` predicate to extension controllers
Extension controllers ought to check the extension class when they receive an event. For backwards compatibility and safety reasons, two cases are considered:

- If the class is not set in the extension, but is set to 'shoot' in the controller (predicate), the request must be accepted.
- If the class is set to 'shoot' in the extension, but is not set in the controller (predicate), the request must be accepted. Extension of class 'garden' must not be considered.

Generally, to follow the example of `ManagedResource` handling, the `class` field is planned to be only set in the `garden` use-case. Gardener won't set the `class` for extensions belonging to shoot clusters.