Fix race-condition during KSM cleanup (#10268)
PR #10062 renamed the kube-state-metrics ManagedResouce (MR) to
kube-state-metrics-runtime or kube-state-metrics-seed. Both the old and
new MR contain the same Prometheus ScrapeConfigs. In the case that the
new MR and their contained resources are created before the old MR was
fully deleted, a situation might occur where the ScrapeConfigs are
deleted again and the new MR is unhealthy until the next reconciliation.

This situation can be reliably triggere by adding a delay to the
deletion of the kube-state-metrics MR in the ManagedResource reconciler:

+++ b/pkg/resourcemanager/controller/managedresource/reconciler.go
@@ -89,6 +89,10 @@ func (r *Reconciler) Reconcile(ctx context.Context, req reconcile.Request) (reco
                return reconcile.Result{}, fmt.Errorf("error retrieving object from store: %w", err)
        }

+       if mr.Name == "kube-state-metrics" && mr.DeletionTimestamp != nil {
+               time.Sleep(10 * time.Second)
+       }
+
        if ignore(mr) && mr.DeletionTimestamp == nil {
                log.Info("Skipping reconciliation since ManagedResource is ignored")
                if err := r.updateConditionsForIgnoredManagedResource(ctx, mr); err != nil {

With this commit, we wait for the deletion of the old kube-state-metrics
MR before creating the new MRs to avoid that ScrapeConfigs are deleted
again.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>