Improve the SeedConditionFailing alert when nodes are rolled (#10270)
* Improve the SeedConditionFailing alert when nodes are rolled out

We recently noticed that the SeedConditionFailing alert might become a
bit too noisy during node rollout. During such process, the affected
conditions become false and the alert is triggered. The noise might
appear if the next node is not rolled out immediately but a few minutes
later. During these few minutes, the unhealthy condition might change to
true and resolve the alert. This commit improves this by extending the
query time window from 5m to 8m.

The alert query is also changed to consider how many samples within such
time window are unhealthy. We choose to alert if 50% or more samples
(i.e., 4 out 8 minutes) are so. This way, we expect to reduce some
flapping as well in situations when a condition becomes unhealthy for 1
or 2 minutes only.

count_over_time(..) >= 4 requires to start the test time series with 3
new samples so that the test starts with the 4th sample. In other words,
count_over_time(..) >= 4 is always false for less than 4 samples. Please
note the notation 0x2 is shorthand short 0+0x2: start with 0 then
followed by two 0's (i.e., 0 0 0).

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Remove one sample too much from the test series

Notation 0xN in PromQL unit tests is shorthand for 0+0xN: start with 0
then followed by N additional 0's. Thus, 0xN equals a series of N+1 0's.
This commit adjusts the test series that had an unnecessary trailing
sample because of this.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

---------

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>