Grant get, list and watch permissions to CRD for authenticated users
This change is required so that shoot owners can also deploy their own
kube-state-metrics and use the custom resource state feature to generate
metrics about their shoots. Such feature requires that the user running
kube-state-metrics has list and watch permissions for the
`customresourcedefinitions` resource in the virtual cluster. These
permissions are required to start kube-state-metrics when a custom
resource state configuration is given, even if shoots do not have a
custom resource definition but are part of the kubernetes API
aggregation layer. The get permission is granted additionally because it
might be useful for further use cases.

Shoot owners will need to run kube-state-metrics with the --namespaces
flag because they are only able to access shoots of their own. If this
flag is not given, kube-state-metrics will log access errors to
forbidden namespaces.

Co-authored-by: Christoph Kleineweber <c.kleineweber@sap.com@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Peter Sutter <peter.sutter@sap.com>