Wait for `Seed` readiness after deploying `Gardenlet`
We should wait for the `Seed` object to be ready after we deployed the
`gardenlet` (via the `seedmanagement.gardener.cloud/v1alpha1.Gardenlet`
resource).

This way, we make sure that the deployment worked before running e2e
tests. In addition, this makes the scenarios consistent (we also wait
for `Seed` readiness when deploying the `gardenlet` Helm chart
directly).