Reduce kublet http2 timeouts. (#10223)
* Reduce kublet http2 timeouts.

In the event that the connection from kubelet to kube-apiserver is silently dropped, it might take 45s to establish a new connection due to the default http2 timeouts.
During this period, node leases are not renewed. On the other hand, the default for node-monitor-grace period in kube-controller-manager is 40s.
After this period, the condition of endpoints in endpointslices will become unready and won't receive any traffic.
This could result in service disruption of user workload following short connection issues from kubelet to kube-apiserver.
By reducing the sum of http timeouts to below 40s, this scenario can be prevented.

* Calculate http2 timeouts depending on nodeMonitorGracePeriod.

* Address review feedback.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

---------

Co-authored-by: Johannes Scheerer <johannes.scheerer@sap.com>
Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>