Restrict `provider-local`'s `Ingress` controller to shoot namespaces
Otherwise, it creates `DNSRecord`s even in the `garden` namespace. When
a seed is the garden cluster at the same time, this leads to issues on
seed deletion because some `DNSRecord`s are still left (those managed by
`gardener-operator` belonging to the garden cluster).

As the purpose of this controller was to just make the domains
resolvable in the shoot machine pods, we now restrict it to the shoot
namespaces.