Fix managed resource migration during KSM upgrade
The shoot-core-kube-state-metrics managed resource used to manage a
cluster role and cluster role binding in the shoot. These are required
so that the kube-state-metrics deployed in the shoot's control-plane
namespace in the seed can collect metrics about the kube-system
namespace in the shoot. This managed resource used to be of class shoot.

During the kube-state-metrics upgrade, the code was migrated so that
such cluster role and cluster role binding are now managed by a new
shoot-core-kube-state-metrics-target managed resource. The old
shoot-core-kube-state-metrics is reused to manage other resources that
are deployed in the shoot's control-plane namespace in the seed. Because
of this, its class changed from shoot to seed. Therefore, during the
migration, the new shoot-core-kube-state-metrics is applied but, because
of the class change, it tries to delete the cluster role and cluster
role binding in the seed cluster instead.

This commit fixes this by removing the shoot-core-kube-state-metrics
managed resource if it's still of class shoot before the migration takes
place.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>