Prevent reconciliation errors in hibernated shoots while migrating KSM
When the shoot is hibernated, the reconciliation fails because the
shoot's API server is not running and the deletion of the managed
resource does not finish. The fix or hack is to remove the finalizer of
the old managed resource and just delete it. This way, the GRM does not
need to run to be able to delete the managed resource, so it works also
on hibernated shoots. Note that the artifacts of the managed resource do
not need to be deleted anyway. We are just migrating them to a different
managed resource with a different name to have a consistent naming for
the managed resources:

- shoot-core-kube-state-metrics for the control-plane in the seed.
- shoot-core-kube-state-metrics-target for the shoot.

Previously, the shoot-core-kube-state-metrics was targeting the shoot.
It seems that repurposing a managed resource that used to target the
shoot to now target the control-plane in the seed has some caveats we
did not anticipate.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Plamen Kokanov <plamen.kokanov@sap.com>