Add alert for VPA capped recommendations for the shoot
This alert is also deployed twofold. First, within the shoot Prometheus,
which in turn is configured to forward alerts to the seed Alertmanager.
The shoot Alertmanager does not play a role here because this alert is
not shared with shoot owners (i.e., the "visibility = operator" label).
Second, the alert is also federated into the garden Prometheus for
further notifications.

The namespace label is used to distinguish between the resources
deployed in the shoot kube-system namespace or in the control-plane
namespace in the seed. It has the "kube-system" value for the former
case and "shoot-control-plane" for the latter.