Add alert for VPA capped recommendations for the seed
This alert is deployed twofold. First, within the aggregate Prometheus,
which in turn is configured to forward alerts to the seed Alertmanager.
Second, this alert is also federated from the aggregate into the garden
Prometheus. The former alert is routed via the default
`email-kubernetes-ops` route, while the latter allows for further
notifications from the garden Prometheus.