Add Garden resource metrics to the runtime kube-state-metrics (#10393)
* Add Garden resource metrics to the runtime kube-state-metrics

bebad97 upgraded kube-state-metrics to a version that supports custom
resource state metrics, and 1e3de9d introduced further enhancements to the
garden-runtime level kube-state-metrics deployment.

This commit makes use of both of these enhancement to introduce new metrics
based on custom resource state metrics. Read more here:
https://github.com/kubernetes/kube-state-metrics/blob/main/docs/metrics/extend/customresourcestate-metrics.md

The Garden resource exposes conditions and information on the last
operation, which can be exposed using a StateSet in kube-state-metric.

* Adapt tests for custom resource metrics

* Remove the shoot test's goroutine modifying the ManagedResource

Commit d4cd212 changed the logic around ManagedResource creation and
removed the check that waits for the MR to become healthy, which eliminates
the need for the goroutine in the shoot test mocking a deployment becoming
healthy.

* Adapt the kube-state-metrics-runtime ClusterRole

* Address review comments