Prevent setting `net.netfilter.nf_conntrack_max` in case `kube-proxy` is deployed. (#10445)
Previously, the kernel setting `net.netfilter.nf_conntrack_max` was
always set during node initialisation. In case `kube-proxy` was deployed
to the nodes it also set the same kernel setting. However, the static
initialisation by Gardener used a static value, but `kube-proxy`
calculated a value based on the number cpu cores.
To reduce the permissions of `kube-proxy`, an init container was
introduced, which set the kernel settings. However, the ordinary
container also tries to do the same thing in case the value is not as
expected. This can lead to an endless crash loop. In case the node
initialisation is performed multiple times as can happen with gardener
node agent.
A simple solution to the problem is to only set kernel setting
`net.netfilter.nf_conntrack_max` in the node initialisation if
`kube-proxy` is not enabled.

Co-authored-by: Johannes Scheerer <johannes.scheerer@sap.com>