Allow disablement of VPA for VPN components.
In some scenarios, the availability of the VPN is crucial and
disruptions due to autoscaling are not desired. This change allows to
disable the autoscaling of the VPN components in shoot/seed per cluster.

This is not meant as a general solution and will never be the standard
for various reasons. However, in certain scenarios it may help to be
able to disable autoscaling to prevent undesired disruptions.