Add Garden condition alert and accompanying dashboard
d7b9f90 introduced new metrics tracking the last operation and conditions.
This commit introduces a new alert if the Garden resource reports
conditions that are not `True` for 10 minutes. It also includes an alert
that fires if the garden-runtime `kube-state-metrics` can't be scraped for
`10m` to avoid a downtime of `kube-state-metrics` suppressing alerts about
conditions.
A supplemental dashboard is included that displays the state of the Garden
resource's conditions and the last operation over time.