Fix setting default namespace when deleting objects with `Applier`
Before, `Applier` was setting a default namespace even for non-namespaced objects. The kube-apiserver does not seem to care, but the fake client for unit tests does. Thus, it was not possible to delete CRDs in unit tests using the `Applier`.
As there is already a logic implemented for the apply case, it is reused for the delete case now.