[operator] Deploy extension via `operator.gardener.cloud/v1alpha1.Extension` resource (#10518)
* Implement runtime deployment

Implement the main logic of the deployment of the extension in the runtime cluster.

* Add new `--controllers` flag

Add `--controllers` flag to enable only an individual set of controllers.

* Adjust `provider-local` extension

* Deploy relevant extension CRDs to runtime cluster

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Refine `extension` reconciliation flow

Refine `extension` reconciliation flow to eliminate
certain shortcuts there were taken previously, e.g.
incomplete deletion logic for `extension`s and the
`garden` cluster.

* Let `gardener-operator` wait for virtual MRs

* Use `ManagedResource` to register extension

Use a `ManagedResource` to manage the `ControllerRegistration` and `ControllerDeployment` to simplify and improve the deletion process of the garden cluster. The `garden` controller then only needs to wait until (virtual) MRs are gone, before it can proceed with deletion.

* Adjust tests

* Update `skaffold-operator.yaml`

* Update documentation

* Pass `priorityClassName` to runtime deployment

* Fix setting default namespace when deleting objects with `Applier`

Before, `Applier` was setting a default namespace even for non-namespaced objects. The kube-apiserver does not seem to care, but the fake client for unit tests does. Thus, it was not possible to delete CRDs in unit tests using the `Applier`.
As there is already a logic implemented for the apply case, it is reused for the delete case now.

* Address PR review feedback v1

* Address PR review feedback v2

* Deploy extension in runtime cluster conditionally when a `BackupBucket`, `DNSRecord` or `Extension` for its type exists

* Address PR review feedback v3

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>
Co-authored-by: oliver-goetz <o.goetz@sap.com>