Fix extracting the go binaries in the remote local setup
Commit 71fc566 reverted the installation of the go binaries back to
commit 583ce9a to download go via its download site. This was the
original behaviour of the remote local setup but it was changed in
commit e80591f to use apk instead.

However, the installation as in commit 583ce9a is not fully correct
because extracting the binary does not consider when a newer go version
is downloaded. In such case, the following extraction command fails
because the wildcard matches both the older and newer go binaries. Note
the downloaded binaries are never removed.

```
tar -C /usr/local -xzf go1.*.linux-amd64.tar.gz
```

This is fixed in this commit by adapting the command above to only
extract the last downloaded version.