Backdate generated certificates slightly (1 min) to allow a certain clock skew of clients. (#10603)
* Backdate generated certificates slightly (1 min) to allow a certain
clock skew of clients.

In the real worlds, clocks are not always in sync. Clock skew is real
and may cause trouble in a distributed system.
For example, if a user creates a kubeconfig via the garden cluster it
should be valid also for the shoot cluster's api server, which may run
on a different infrastructure and hence may use different time servers
for synchronisation.
Backdating the certificates slightly allows some clock skew and prevents
confusing errors.

* Address review feedback.

Co-authored-by: Viktor <v.kostov@sap.com>

---------

Co-authored-by: Viktor <v.kostov@sap.com>