Fix cAdvisor network metric relabeling config
The cAdvisor relabeling config uses the name of the network interface to
add a label to the scraped metrics if it is a node's host network. This
happens for different devices, i.e. if overlay network is enabled, it also
applies to the overlay network device (tunl0). Critically, nodes on AWS
will name the device as specified here:
https://github.com/systemd/systemd/blob/ccddd104fc95e0e769142af6e1fe1edec5be70a6/src/udev/udev-builtin-net_id.c#L29
This leads to device names such as `enp39s0`, which were, prior to this
commit, not matched by the regex.
In consequence, the `Node Details` dashboard's network throughput panel did
not display data, as it filters for the metrics labeled
`host_network="true"`.
So far, this was only an issue on AWS nodes. OpenStack, Aliyun, and GCP
name devices `ens.+`, and Azure node network devices are named `eth0`.
Therefore, adapting the regex in the relabling config to match ethernet
devices (`en` prefix) will also populate the `Node Details` dashboard's
network panel again.