Adapt sequence of garden deletion
The `Extension` controller deploys admission charts via managed resources in the virtual cluster.
The current delete flow of a `Garden` can get stuck if an `Extension` deploys an admission chart.
When we adapt the sequence of the deletion flow, we can improve the situation.

1. Delete MR of `kube-controller-manager` and `gardener-apiserver` after the other MRs for the virtual cluster have been deleted.
`kube-controller-manager` is required to clean-up OwnerReferences. If it is deleted too early and there are OwnerReferences which belong to resources created by MRs left, the Garden deletion is stuck.
`gardener-apiserver` is required to clean-up Gardener objects. If it is delete too early and an Extension uses a Gardener resource in its application helm chart, the Garden deletion is stuck.

2. Delete MR of `gardeneraccess` and `istio` later in the process
`gardeneraccess` MR is now deleted together with the components of 1. This simplifies debugging and manual cleanup, because the user does not lose access to virtual garden until `Extension` MRs are deleted.
`istio` is now deleted after virtual-garden was deleted. Otherwise, the user looses access to the virtual-garden when it is still running and is not able to unblock the deletion manually.