Propagate `Infrastructure.status.egressCIDRs` to `Shoot.status.networking.egressCIDRs` (#10240)
* Add `Shoot.status.networking.egress`

* Add validation for new field

* Propagate egress CIDRs from `Infrastructure` status to `Shoot` status

This commit changes the logic in `WaitForInfrastructure` bit.
Previously, `shoot.Status.Networking` was only updated if all 3 of nodes, services, and pods CIDR have been reported in the
`Infrastructure` status. Now, we always update `shoot.Status.Networking` with what has been reported in the `Infrastructure`.
With this, we always sync the `Cluster` resource after this operation as well.
This should be more like what we want: if any of the CIDRs doesn't exist in the `Infrastructure`, we still want to propagate the
others back to the `Shoot` status.

* Report egress IPs in provider-local

* Add check for reported CIDRs in e2e tests

* Rename status field to `egressCIDRs`

* Fix e2e tests for workerless clusters