[operator] Cleanup webhook configuration from virtual cluster when removing admission deployment (#10585)
* Introduce `webhook-config-owner-namespace` CLI option to extensions library

* Create own `extension-*` namespace in virtual cluster and use it as owner reference for webhook configuration

* Address PR review feedback

* Address PR review feedback v2

* Address PR review feedback v3

* Adapt sequence of garden deletion

The `Extension` controller deploys admission charts via managed resources in the virtual cluster.
The current delete flow of a `Garden` can get stuck if an `Extension` deploys an admission chart.
When we adapt the sequence of the deletion flow, we can improve the situation.

1. Delete MR of `kube-controller-manager` and `gardener-apiserver` after the other MRs for the virtual cluster have been deleted.
`kube-controller-manager` is required to clean-up OwnerReferences. If it is deleted too early and there are OwnerReferences which belong to resources created by MRs left, the Garden deletion is stuck.
`gardener-apiserver` is required to clean-up Gardener objects. If it is delete too early and an Extension uses a Gardener resource in its application helm chart, the Garden deletion is stuck.

2. Delete MR of `gardeneraccess` and `istio` later in the process
`gardeneraccess` MR is now deleted together with the components of 1. This simplifies debugging and manual cleanup, because the user does not lose access to virtual garden until `Extension` MRs are deleted.
`istio` is now deleted after virtual-garden was deleted. Otherwise, the user looses access to the virtual-garden when it is still running and is not able to unblock the deletion manually.

* Address PR review feedback v4

* Undo move deletion of `gardener-apiserver` to a later point

Moving the deletion of `gardener-apiserver` seems to introduce additional dependecies to `gardener-controller-manager` e.g. for removing finalizers from `ControllerRegistrations`.
Instead of postponing more and more components `gardener-apiserver` is now deleted on its original place again. GRM uses `IgnoreNotFound` which might also cover the error when the whole object is not available anymore.

* Address PR review feedback v5