[operator] Add extension `Required` controller (#10650)
* Moving main controller to separate `extension` package

* Refactor: Make fields private

* Enhance `OperatorConfiguration` API

Enhance `OperatorConfiguration` API and ass missing defaults for the configuration of the extension controller.

* Add `required` controller

Move the `required` calculation from the main controller to a dedicated one. It works with a new `RequiredRuntime` condition for `extension` resources, which the main controller processes. Later, this controller will be the used to calculate a second `require` condition, based on available `ControllerInstallation`s.

* Refine extension deletion flow

Make the extension deletion flow aware of the new required condition.

* Enhance documentation

* Update Skaffold dependencies

* Address review comments

* Add log entries for troubleshooting

* Merge status with optimistic locking

* Ignore not found errors

The `BackupBucket` controller complains about a missing secret during e2e tests:

Secret \"test-backup-bucket\" not found"

While this is very much related to the e2e setup, it's still reasonable to continue with the deletion, if the secret is already gone.

* Address review comment

* Address review comment