Support for Structured Authorization for `Shoot`s >= Kubernetes v1.30 (#10682)
* Documentation

* Augment API

* [make generate]

* Augment seed authorizer graph logic

according to new API

* Add validation handler to GAC for authz configs

Similar to how its done for audit policies or authn configs

* Adapt `kubernetes/apiserver` component to support structured authz

If the cluster is < 1.30 or the feature gate is disabled, the old method
of configuration authz settings is still available.
However, for 1.30+ clusters, always the new structured authz approach is
used.

* Add condition to only intercept requests from gardenlets and service accounts from seed namespaces

This only works for 1.30+ clusters using the structured authz approach,
but it's a nice feature that we can use now (with < 1.30, it was not
possible to configure such conditions).

Also, this allows us to set the failure policy to "deny" (i.e., when our
webhook server denies a request from a gardenlet or an extension, nobody
else should be able to allow it (seed authorizers is the only instance
which may decide (except for manually configured RBAC rules)).

* Adapt config computation for `kubeapiserver` component

to read the referenced authz ConfigMap and compute the proper config for
the deployer

* Rename `admissionpluginsecret` to `shootkubeconfigsecretref`

This admission handler ensures that a kubeconfig is not removed from
Secrets in case it is referenced in shoots.

In the next commit, we augment it with verification for authorization
kubeconfigs.

* Augment shoot kubeconfig validator with authz kubeconfig secret handling

* Address PR review feedback

* [PR feedback] Adapt `reference` controllers

* [PR feedback] Ensure shoot references exist

* Address PR review feedback

* Address PR review feedback

* Add TODO