[GEP-28] Initial `gardenadm` skaffolding and introduction of new component (#10729)
* Move `InitRun` function into isolated package

This file has been deliberately moved into its own package since it imports k8s.io/component-base/version/verflag
which automatically registers the `--version` flag as soon as the packages is (transitively) imported.
In order to prevent this from happening accidentally, it's safer to keep it in its own package.

* Boilerplate for new `gardenadm` binary

* Address PR review feedback

* [PR review feedback] Exit on error

* Address PR review feedback