Increase the readiness probe timeout for the gardener-metrics-exporter (#10771)
The gardener-metrics-exporter readiness probe targets the "/metrics"
endpoint. This, however, becomes slower as more shoots and metrics need
to be exported. Eventually, this might take longer than the default
timeout of 1s.

The liveness probe, on the other hand, only probes the "/" endpoint,
which is very fast and does not get slower.

This commit increases the readiness probe timeout from 1s to 10s and the
probe period from 5s to 15s.

Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>