[operator] Deploy `BackupBucket`/`DNSRecord` extension resources (#10645)
* `Garden` controller deploys `extensions.gardener.cloud/v1alpha1.BackupBucket` resource

Only delete extension after all extension resources have been cleaned up

remove AnnotationKeyExtensionResourcesCleanedUp annotation

* `DNSRecord` deployer only creates `Secret` when data is provided

In the garden case, the DNS secret is already part of the `garden` namespace (where the `DNSRecord`s) are created, so there is no need to create another new `Secret` and replicate the credentials into it. We can just reference it.

* `Garden` controller deploys `extensions.gardener.cloud/v1alpha1.DNSRecord` resources

* `garden-reference` controller checks DNS secret ref

* address review feedback

* Support separate DNS providers per domain name.

The section `.spec.dns` is introduced for specifying the DNS providers. The `domains` fields are objects instead of strings. If the field `provider` is set, the provider is looked up in the DNS section.

* address second review

* make generate

* fix required controllers to consider extension class for requiredTypes

* only wait for provider-local extension deletion after garden is deleted

* address review

* address review feedback

* fix backupbucket container/prefix calculation

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>