Refine `gardener-scheduler` predicates
The scheduler should only react to update events when a shoot is not assigned yet and the spec has changed. If the shoot is not assignable at the moment, the reconciliation backoff of the controller leads to an automatic re-enqueuing of the shoot.
Earlier, the controller reacted to shoot status update events which resulted in frequent reconciliations.