Allow control-plane blackbox exporter to talk to kube-apiserver service
The blackbox exporter uses the `kube-apiserver` service to verify that the
VPN works. It does this by requesting logs for a container running in the
shoot cluster (vpn-shoot-0's init container). When the network policy label
above was removed in 413a66c (g/g#10775), blackbox-exporter requests to the
API server time out. This, in turn, means that the `probe_success` metric
for this job in Prometheus becomes 0, and the VPN dashboard to show that
the VPN is down, when it is, in fact. reachable.