[node-agent-authorizer] Introduce webhook based authorization to `gardener-node-agent` (#10781)
* Introduce `NodeAgentAuthorizer` feature gate

* Make CSR certificate request reusable

* Support certificate based authentication in `gardener-node-agent`

* Enable `NodeAgentAuthorizer` authorization webhook in shoot kube-apiservers

* Activate `NodeAgentAuthorizer` feature gate for local setup

* Check if node exists before denying gardener-node-agent CSR request

A `machine` can have a `node` label even before kubelet created it.

* Start lease controller if node name is not empty only

Restart gardener-node-agent when node was registered by kubelet.
As a side effect, this ensures that gardener-node-agent watches its own node only.

* Create integration test for `RequestAndStoreKubeconfig`

You can't get a proper rest.Config for a fake client, so we need an integration test to test it.

* Address PR review feedback

* Address PR review feedback v2

* Refactor `app.go`

* gardener-node-agent fetches machine-name from if NodeAgentAuthorizer feature gate is active only

If the feature gate is not active, the machine-name is not required and the corresponding file is not created.