Adapt operator config API
The operator config API reuses the `ETCDConfig` type from the gardenlet config API.
gardener-operator passes these config values when instantiating the etcd and etcd-druid components.

This commit switches the internal operator config API version to also use the external version of the
gardenlet `ETCDConfig` type. With this, we spare the temporary conversion in gardener-operator until
we drop the internal version of the operator config API.
The overall goal is to tackle both components in dedicated PRs to keep them reviewable.