Drop internal version of `resourcemanager.config.gardener.cloud` API (#11099)
* Move API validation

The validation is now implemented on the external version of the config API that includes the external version of the
`componentbaseconfig.{LeaderElection,ClientConnection}Configuration` types.
However, the validation for these types is implemented on the internal version.
Hence, this commit also adds conversion for these types from external to internal so that we can continue using the
provided componentbaseconfig validation code.

* Drop internal API version

* Adapt entrypoint

* Adapt imports

* Update skaffold dependencies