Allow IPv6-only tests in test machinery to create shoot cluster.
On real infrastructures, we assume that the infrastructure will provide
the IPv6 ranges. Therefore, no node IP range should be specified for
those tests.
To keep the changes minimal, the default IPv4 range is kept in place,
but it is removed for IPv6-only clusters so that the cluster creation
succeeds.