Drop internal version of `gardenlet.config.gardener.cloud` API (#11101)
* Prefactor: remove indirection of token requestor controllers

* Move API validation

The gardenlet config validation calls the validation for the `SeedTemplate` on the core API's internal version.
Hence, we add a helper that allows us convert the external `SeedTemplate` in the external gardenlet config API
to the internal core representation where the validation is implemented.

* Move API helper

* Drop internal API version

* Adapt conversion

Some code previously converted an external gardenlet config to the internal representation (e.g., `ManagedSeed{,Set}` validation,
`ManagedSeed` admission, gardenlet deployer). These occurrences are adapted now to handle the external gardenlet config directly
and convert only the included `SeedTemplate` to the internal version if needed (e.g., for validation, defaulting).

* Adapt entrypoint

* Adapt imports

* Adapt operator config API

The operator config API reuses the `ETCDConfig` type from the gardenlet config API.
gardener-operator passes these config values when instantiating the etcd and etcd-druid components.

This commit switches the internal operator config API version to also use the external version of the
gardenlet `ETCDConfig` type. With this, we spare the temporary conversion in gardener-operator until
we drop the internal version of the operator config API.
The overall goal is to tackle both components in dedicated PRs to keep them reviewable.

* Update skaffold dependencies

* Address review suggestions