Restrict `gardenlet` access for `Seed` resources (#11062)
* No longer unconditionally allow `update`, `patch` verbs for `Seed`s

The `gardenlet` only updates its own `Seed`, so there is no need to have
these verbs as part of the "always allowed" list.

* No longer unconditionally allow `delete` verb for `Seed`s

The only place where `gardenlet` deletes `Seed` resources is in its
`ManagedSeed` controller via
https://github.com/gardener/gardener/blob/48169a21cfbfd5a515a7374c729c9de9111e5bd2/pkg/controller/gardenletdeployer/actuator.go#L304-L311.

Now, the graph is consulted and asked whether the `gardenlet` is allowed
to delete it.

* New controller in GCM writing `ReadyForMigration` shoot constraint

* Maintain seed labels on `Seed` resource

- own name
- optional: name of parent seed in case the `Seed` is a managed seed

This allows to use a label selector in `gardenlet` when watching
`Seed`s.

Similar to https://github.com/gardener/gardener/pull/9089

* Maintain seed label on `ManagedSeed` resource

only the name of the parent seed (i.e., the name of the seed that is
used in the referenced shoot)

* Add TODO for restricting `gardenlet` watches on `{Managed}Seed`s with label selector

This should not be activated in the same release since we have to
make sure that all `{Managed}Seed`s get the label first (which usually
only happens when the `gardenlet`s make their first request for updating
the `{Managed}Seed` status).

If we rolled this out in one release (gardenlet cache already
restricted), the `{Managed}Seed`s wouldn't have the labels yet, hence
the `gardenlet`s wouldn't see any `{Managed}Seed`s.

* Improve `migration` controller

The controller is ought to remove the `ShootReadyForMigration`
constraint in case it missed the event during migration.

* Remove constraint only during restoration

The `ReadyForMigration` taint was previously removed at the beginning of the `migration` phase.
However, if the migration reconcile flow restarts then gardenlet will end up again and stuck in the check ready-for-migration check (see https://github.com/gardener/gardener/blob/4c6e4d23d2725c93ac77cae246095e20af75d86a/pkg/gardenlet/controller/shoot/shoot/reconciler.go#L161).
With this change the constraint is removed during the `restore` phase.

* Change to external config

* Address review comments

---------

Co-authored-by: Tim Usner <tim.usner@sap.com>