First steps towards local dual-stack setup (#11226)
* Switch local dual-stack setup to IPv6 as primary address family.

* Ensure istio is listening on both address families in case of dual-stack

* Adapt `kube-apiserver` service to dual-stack seed clusters.

In case the seed cluster is dual-stack, the kube-apiserver service
should also be dual-stack. This provides an address in the correct
address family regardless of the shoot cluster's address families.
For example, in an IPv6/IPv4 seed cluster it would be impossible to use
the kube-apiserver primary service IP address, which would be IPv6, in
a single-stack IPv4 shoot cluster for the apiserver-proxy. With a
dual-stack service, it is possible to use the IPv4 address instead.