Fix the alert query for a race condition
The original query assumes that querying Prometheus for the VPA state at
a certain time always yields a result consistent with the actual state
at that specific time. However, we found the KSM metrics are refreshed
independently of each other. A race condition can occur after a VPA
recommendation update if the Prometheus scrape happens at an instant
when the uncapped target has already been refreshed but the regular
target has not.

The proposed fix is to use the `count_over_time()` function to ensure
the uncapped target is larger than the regular target for 5m, and alert
only in such case. This is similar to adding `for: 5m` in the alert
definition, but it avoids the pending state, which would be incorrect.