Do not count the VPAs for the garden cluster
Counting the VPAs in the garden cluster has the caveat that labels are
lost during aggregation. This, in turn, complicates debugging which VPA
and containers are affected. To keep the count and the labels, we would
need two queries, similarly as in the seed and shoot cases (i.e., the
aggregating query in those cases is the federated one). We opt to have
only one query for the sake of simplification and an alert per VPA in
the garden cluster.